Upgraded to Pyton 3.5
This version will no longer work with anything below 3.0. It used a
different queue module to keep up with naming conventions (although
that will have be fixed for the rest of the program).

Note: when running the program in decentralized mode, it is much slower
than centralized. This could be due to running almost twice as many
threads, however, it can also be because of the code. Will have to
investigate further.